Update zsh prompt
The git branch logic slowed down the terminal quite a bit because I was
using the "git_prompt_info" zsh command just to get the branch name.
This was previously used to get the git status of the repo, which is
helpful to see changes, but not at the cost of a slow terminal.

Instead of the full info, the prompt now just retrieves the branch name
and if we're in a git repo, then it checks the status for via git status
--porcelain to change the color of the branch name. This increases the
response time dramatically.