vim: Stop lazy loading nvim-tree
The keymapping <leader>k won't work on first start because nvim-tree is
not enabled until one of NvimTreeFocus or NvimTreeToggle is invoked.
This hurts the usability. Alternatively, we could wrap the keymap to
require nvim-tree to be loaded first, but unless there are perf issues,
we should just keep it enabled.